too many changes to list, many regressions were made, master will be UNSTABLE for a few days

+++ src/otclient/net/protocolgameparse.cpp
@@ -251,7 +251,7 @@ void ProtocolGame::parseMessage(InputMessage& msg)
             //case Otc::GameServerObjectInfo:
             //case Otc::GameServerPlayerInventory:
             default:
-                Fw::throwException("unknown opt byte ", opt);
+                Fw::throwException("unknown opt byte ", (int)opt);
                 break;
             }
         }
